Fixed Minigame and Puzzle Crashes on x64 builds

+++ Lawn/SeedPacket.cpp
@@ -67,7 +67,7 @@ void SeedPacket::PickNextSlotMachineSeed()
 			}
 		}
 
-		aSeedWeightArray[aSeedsCount].mItem = (int)aSeedType;
+		aSeedWeightArray[aSeedsCount].mItem = aSeedType;
 		aSeedWeightArray[aSeedsCount].mWeight = aWeight;
 		aSeedsCount++;
 	}
